[HD] Move mnemonic wordlists to separate directory
